+++ detection/api/include/frame_transformers/FrameRotator.h
@@ -42,19 +42,19 @@ namespace MPF { namespace COMPONENT {
     public:
         FrameRotator(IFrameTransformer::Ptr innerTransform, int rotationDegrees);
 
-        cv::Size GetFrameSize(int frameIndex) override;
+        cv::Size GetFrameSize(long frameIndex) override;
 
 
     protected:
-        void DoFrameTransform(cv::Mat &frame, int frameIndex) override;
+        void DoFrameTransform(cv::Mat &frame, long frameIndex) override;
 
-        void DoReverseTransform(MPFImageLocation &imageLocation, int frameIndex) override;
+        void DoReverseTransform(MPFImageLocation &imageLocation, long frameIndex) override;
 
 
     private:
         const int rotationDegrees_;
 
-        cv::Point GetRevertedTopLeftCorner(const MPFImageLocation &imageLocation, int frameIndex) const;
+        cv::Point GetRevertedTopLeftCorner(const MPFImageLocation &imageLocation, long frameIndex) const;
     };
 }}
