+++ detection/api/include/frame_transformers/IFrameTransformer.h
@@ -45,11 +45,11 @@ namespace MPF { namespace COMPONENT {
 
         virtual ~IFrameTransformer() = default;
 
-        virtual void TransformFrame(cv::Mat &frame, int frameIndex) = 0;
+        virtual void TransformFrame(cv::Mat &frame, long frameIndex) = 0;
 
-        virtual void ReverseTransform(MPFImageLocation &imageLocation, int frameIndex) = 0;
+        virtual void ReverseTransform(MPFImageLocation &imageLocation, long frameIndex) = 0;
 
-        virtual cv::Size GetFrameSize(int frameIndex) = 0;
+        virtual cv::Size GetFrameSize(long frameIndex) = 0;
     };
 }}
