+++ detection/api/src/FrameFilter.cpp
@@ -31,9 +31,9 @@
 
 namespace MPF { namespace COMPONENT {
 
-    bool FrameFilter::IsPastEndOfSegment(int originalPosition) const {
-        int lastSegmentPos = GetSegmentFrameCount() - 1;
-        int lastOriginalPos = SegmentToOriginalFramePosition(lastSegmentPos);
+    bool FrameFilter::IsPastEndOfSegment(long originalPosition) const {
+        long lastSegmentPos = GetSegmentFrameCount() - 1;
+        long lastOriginalPos = SegmentToOriginalFramePosition(lastSegmentPos);
         return originalPosition > lastOriginalPos;
     }
 
@@ -43,30 +43,30 @@ namespace MPF { namespace COMPONENT {
     }
 
 
-    double FrameFilter::GetCurrentSegmentTimeInMillis(int originalPosition, double originalFrameRate) const {
-        int segmentPos = OriginalToSegmentFramePosition(originalPosition);
+    double FrameFilter::GetCurrentSegmentTimeInMillis(long originalPosition, double originalFrameRate) const {
+        long segmentPos = OriginalToSegmentFramePosition(originalPosition);
         double framesPerSecond = GetSegmentFrameRate(originalFrameRate);
         double timeInSeconds = segmentPos / framesPerSecond;
         return timeInSeconds * 1000;
     }
 
 
-    int FrameFilter::MillisToSegmentFramePosition(double originalFrameRate, double segmentMilliseconds) const {
+    long FrameFilter::MillisToSegmentFramePosition(double originalFrameRate, double segmentMilliseconds) const {
         double segmentFps = GetSegmentFrameRate(originalFrameRate);
-        return static_cast<int>(segmentFps * segmentMilliseconds / 1000);
+        return static_cast<long>(segmentFps * segmentMilliseconds / 1000);
     }
 
-    double FrameFilter::GetSegmentFramePositionRatio(int originalPosition) const {
+    double FrameFilter::GetSegmentFramePositionRatio(long originalPosition) const {
         double segmentPosition = OriginalToSegmentFramePosition(originalPosition);
         return segmentPosition / GetSegmentFrameCount();
     }
 
-    int FrameFilter::RatioToOriginalFramePosition(double ratio) const {
-        auto segmentPosition = static_cast<int>(GetSegmentFrameCount() * ratio);
+    long FrameFilter::RatioToOriginalFramePosition(double ratio) const {
+        auto segmentPosition = static_cast<long>(GetSegmentFrameCount() * ratio);
         return SegmentToOriginalFramePosition(segmentPosition);
     }
 
-    FrameFilter::CPtr FrameFilter::GetNoOpFilter(int frameCount) {
+    FrameFilter::CPtr FrameFilter::GetNoOpFilter(long frameCount) {
         return CPtr(new IntervalFrameFilter(0, frameCount - 1, 1));
     }
